Create Snarebriars Game
This was the final project on my CodeNation Bootcamp. It's a fully playable text based adventure RPG including ascii artwork and use of timing code to create narrative effects. It was a group project between four of us, but we all did a huge amount of work on it over the three days we had to code it. It was by far the most complex and longest game submitted in my class.